Fix implicit 64-to-32 bit integer narrowing
Summary:
These fire when a 64-bit integer (e.g., `size_t`, `int64_t`) is implicitly converted to a 32-bit type (e.g., `int`, `uint32_t`), which can cause data loss.

Fixes applied:
- Widened variable types to match source (e.g., `int` → `size_t`)
- Added explicit `static_cast` where narrowing is intentional and safe

Differential Revision: D97079855

+++ momentum/marker_tracking/tracker_utils.cpp
@@ -307,7 +307,9 @@ momentum::Character locatorsToSkinnedLocators(
 
     // Set the locator's parent to the closest face.
     std::tie(skinnedLocator.parents, skinnedLocator.skinWeights) = averageTriangleSkinWeights(
-        sourceCharacter, closestPointResult.triangleIdx, closestPointResult.baryCoords);
+        sourceCharacter,
+        gsl::narrow_cast<int>(closestPointResult.triangleIdx),
+        closestPointResult.baryCoords);
     skinnedLocators.push_back(skinnedLocator);
   }
 
@@ -582,7 +584,7 @@ std::vector<momentum::SkinnedLocatorTriangleConstraintT<float>> createSkinnedLoc
     }
 
     momentum::SkinnedLocatorTriangleConstraintT<float> constr;
-    constr.locatorIndex = i;
+    constr.locatorIndex = gsl::narrow_cast<int>(i);
     constr.tgtTriangleIndices = mesh.faces[closestPointResult.triangleIdx];
     constr.tgtTriangleBaryCoords = closestPointResult.baryCoords;
     constr.depth = targetDepth;
